% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyAttributesMapping.R
\name{ApplyAttributesMapping}
\alias{ApplyAttributesMapping}
\title{ApplyAttributesMapping}
\usage{
ApplyAttributesMapping(originalData, attrMapping)
}
\arguments{
\item{originalData}{Original data. Required.}

\item{attrMapping}{List of attributes mapping between original data and internal representation.
Required.}
}
\value{
data.table
}
\description{
Applies attributes mapping to original data
}
\examples{
\dontrun{
ApplyAttributesMapping(originalData, attrMapping)
}

}
