% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyOriginGrouping.R
\name{ApplyOriginGrouping}
\alias{ApplyOriginGrouping}
\title{ApplyOriginGrouping}
\usage{
ApplyOriginGrouping(inputData, originGrouping)
}
\arguments{
\item{inputData}{Input data. Required.}

\item{originGrouping}{Data.table object with mapping from RegionOfOrigin to GroupOfOrigin.
Required.}
}
\value{
inputData
}
\description{
Applies RegionOfOrigin grouping map to the input data
}
\examples{
\dontrun{
ApplyOriginGrouping(inputData, originGrouping)
}

}
