% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppServer.R
\name{AppServer}
\alias{AppServer}
\title{AppServer}
\usage{
AppServer(input, output, session)
}
\arguments{
\item{input}{Input}

\item{output}{Output}

\item{session}{Session}
}
\value{
NULL (invisibly)
}
\description{
Application server logic
}
