% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertStringToDate.R
\name{ConvertStringToDate}
\alias{ConvertStringToDate}
\title{ConvertStringToDate}
\usage{
ConvertStringToDate(str)
}
\arguments{
\item{str}{str}
}
\value{
Date
}
\description{
ConvertStringToDate
}
\examples{
ConvertStringToDate(str = c(
  NA_character_, 'UNK', '', '2020', '2020-Q1', '2020-Q02', '2020/Q2', 'Q1-2020', 'Q03-2020',
  '2020/q2', 'q1-2020', 'Q04/2020', '2020-W01', '2020-W45', '2020-w40', 'W02-2020', 'w10-2020',
  '2020-05-20', '2020/1/10', '20/5/2020', '2020-1', '2020-02', '2020/06', '05-2020', '6/2020'
))

}
