% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hivEstimatesAccuracy2-package.R
\docType{data}
\name{countryData}
\alias{countryData}
\title{Country names, codes, regions information}
\format{
data.table
}
\usage{
countryData
}
\description{
A dataset containing country codes in two-letter format
}
\keyword{datasets}
