% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAdjustmentSpecFileNames.R
\name{GetAdjustmentSpecFileNames}
\alias{GetAdjustmentSpecFileNames}
\title{GetAdjustmentSpecFileNames}
\usage{
GetAdjustmentSpecFileNames(path = GetSystemFile("adjustments"))
}
\arguments{
\item{path}{Path to the folder with adjustment specifications. Optional.
Default = \code{\link{GetSystemFile}('adjustments')}.}
}
\value{
Character vector of adjustment specification file paths.
}
\description{
Get the list of adjustment file names in a specific folder. By default the search folder is
\code{adjustments} folder installed with package \code{\link{hivPlatform}}.
}
\examples{
\dontrun{
GetAdjustmentSpecFileNames()
}

}
