% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAdjustmentSpecs.R
\name{GetAdjustmentSpecs}
\alias{GetAdjustmentSpecs}
\title{GetAdjustmentSpecs}
\usage{
GetAdjustmentSpecs(adjustmentNames)
}
\arguments{
\item{adjustmentNames}{adjustmentNames}
}
\value{
list
}
\description{
GetAdjustmentSpecs
}
\examples{
GetAdjustmentSpecs(c('Multiple Imputation using Chained Equations - MICE'))

}
