% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAdjustmentSpecsWithParams.R
\name{GetAdjustmentSpecsWithParams}
\alias{GetAdjustmentSpecsWithParams}
\title{GetAdjustmentSpecsWithParams}
\usage{
GetAdjustmentSpecsWithParams(params)
}
\arguments{
\item{params}{Params}
}
\value{
list
}
\description{
GetAdjustmentSpecsWithParams
}
\examples{
GetAdjustmentSpecs(c('Multiple Imputation using Chained Equations - MICE'))

}
