% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAttrMappingStatus.R
\name{GetAttrMappingStatus}
\alias{GetAttrMappingStatus}
\title{GetAttrMappingStatus}
\usage{
GetAttrMappingStatus(attrMapping)
}
\arguments{
\item{attrMapping}{List of attributes mapping between original data and internal representation.
Required.}
}
\value{
list object
}
\description{
Checks attributes mapping and returns detailed information about errors if encountered.
}
\examples{
\dontrun{
GetAttrMappingStatus(attrMapping)
}

}
