% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBootstrapFitStats.R
\name{GetBootstrapFitStats}
\alias{GetBootstrapFitStats}
\title{GetBootstrapFitStats}
\usage{
GetBootstrapFitStats(fits)
}
\arguments{
\item{fits}{fits}
}
\value{
list
}
\description{
Get bootstrap fits statistics
}
\examples{
\dontrun{
GetBootstrapFitStats(fits)
}

}
