% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDate.R
\name{GetDate}
\alias{GetDate}
\title{GetDate}
\usage{
GetDate(y = NA, q = NA, m = NA, w = NA, d = NA)
}
\arguments{
\item{y}{y}

\item{q}{q}

\item{m}{m}

\item{w}{w}

\item{d}{d}
}
\value{
Date
}
\description{
GetDate
}
\examples{
GetDate(1977, 2, 5, NA, 20)

}
