% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHIVPlotData.R
\name{GetHIVPlotData}
\alias{GetHIVPlotData}
\title{GetHIVPlotData}
\usage{
GetHIVPlotData(mainFitResult, bootstrapFitStats)
}
\arguments{
\item{mainFitResult}{mainFitResult}

\item{bootstrapFitStats}{bootstrapFitStats}
}
\value{
data.table
}
\description{
Get HIV plot data
}
\examples{
\dontrun{
GetHIVPlotData <- function(
  mainFitResult,
  bootstrapFitStats
)
}

}
