% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInputDataValidityStatus.R
\name{GetInputDataValidityStatus}
\alias{GetInputDataValidityStatus}
\title{GetInputDataValidityStatus}
\usage{
GetInputDataValidityStatus(inputData, columnSpecs = NULL)
}
\arguments{
\item{inputData}{Input data. Required.}

\item{columnSpecs}{columnSpecs}
}
\value{
list
}
\description{
Validate input data.
}
\examples{
\dontrun{
GetInputDataValidityStatus(inputData)
}

}
