% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetListObject.R
\name{GetListObject}
\alias{GetListObject}
\title{GetListObject}
\usage{
GetListObject(fileName, section = NULL, includeFileName = TRUE)
}
\arguments{
\item{fileName}{Name of list file. Required.}

\item{section}{Name of section of the list to return. If not specified, then
the whole specification is returned, otherwise only the specified section. Optional.
Default = NULL.}

\item{includeFileName}{Logical indicating to include file name in the returned list as field
"FileName". Optional. Default = TRUE.}
}
\value{
List object.
}
\description{
Returns list object stored as text in R script file.
}
\examples{
\dontrun{
GetListObject(fileName)
GetListObject(fileName, section = 'Name')
GetListObject(fileName, section = 'Parameters')
}

}
