% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMainReportArtifacts.R
\name{GetMainReportArtifacts}
\alias{GetMainReportArtifacts}
\title{GetMainReportArtifacts}
\usage{
GetMainReportArtifacts(params)
}
\arguments{
\item{params}{Main report parameters list object}
}
\value{
List of table and plot objects
}
\description{
Get all tables and plots used for population of the main report.
}
\examples{
\dontrun{
GetMainReportArtifacts(params)
}

}
