% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMCMCAutoCorrelationPlot.R
\name{GetMCMCAutoCorrelationPlot}
\alias{GetMCMCAutoCorrelationPlot}
\title{GetMCMCAutoCorrelationPlot}
\usage{
GetMCMCAutoCorrelationPlot(covariance)
}
\arguments{
\item{covariance}{Array of covariance matrices. Required.}
}
\description{
Get autocorrelation plot from MCMC chain convergence.
}
\examples{
\dontrun{
GetMCMCAutoCorrelationPlot(covariance)
}

}
