% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMCMCBetasPlot.R
\name{GetMCMCBetasPlot}
\alias{GetMCMCBetasPlot}
\title{GetMCMCBetasPlot}
\usage{
GetMCMCBetasPlot(betas)
}
\arguments{
\item{betas}{Array of beta matrices. Required.}
}
\description{
Get plot of Betas from MCMC chain convergence.
}
\examples{
\dontrun{
GetMCMCBetasPlot(betas)
}

}
