% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMiceDiagnosticPlot.R
\name{GetMiceDiagnosticPlot}
\alias{GetMiceDiagnosticPlot}
\title{GetMiceDiagnosticPlot}
\usage{
GetMiceDiagnosticPlot(mids)
}
\arguments{
\item{mids}{Object of class "mids" (output of \code{\link[mice]{mice}}) function.}
}
\description{
Gets mice diagnostic plot.
}
\examples{
\dontrun{
GetMiceDiagnosticPlot(mids)
}

}
