% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMissingnessPlots.R
\name{GetMissingnessPlots}
\alias{GetMissingnessPlots}
\title{GetMissingnessPlots}
\usage{
GetMissingnessPlots(
  inputData,
  columnNames = c("Transmission", "GroupedRegionOfOrigin", "Age", "FirstCD4Count"),
  labels = c("Transmission", "Migrant", "Age", "CD4")
)
}
\arguments{
\item{inputData}{Pre-processed input data. Required.}

\item{columnNames}{Names of columns in \code{inputData} selected for missingness plot. Optional.
Default = \code{c('Transmission', 'GroupedRegionOfOrigin', 'Gender', 'Age', 'FirstCD4Count')}}

\item{labels}{Labels to be displayed instead of strings in \code{columnNames}. Optional.
Default = \code{c('Transm.', 'Migrant', 'Gender', 'Age', 'CD4')}}
}
\value{
list object
}
\description{
Create plot with overview of missing data.
}
\examples{
\dontrun{
GetMissingnessPlots(inputData)
}

}
