% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNiceUpperLimit.R
\name{GetNiceUpperLimit}
\alias{GetNiceUpperLimit}
\title{GetNiceUpperLimit}
\usage{
GetNiceUpperLimit(val)
}
\arguments{
\item{val}{Vector of numeric values. Required.}
}
\value{
numeric nice number
}
\description{
Generates a "nice" large number.
}
\examples{
GetNiceUpperLimit(val = 8)
GetNiceUpperLimit(val = -8)
GetNiceUpperLimit(val = c(8, 13, 1158, -13232))

}
