% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOriginDistribution.R
\name{GetOriginDistribution}
\alias{GetOriginDistribution}
\title{GetOriginDistribution}
\usage{
GetOriginDistribution(inputData)
}
\arguments{
\item{inputData}{Input data. Required.}
}
\value{
data.table
}
\description{
Get distribution of RegionOfOrigin
}
\examples{
inputData <- data.table::data.table(FullRegionOfOrigin = c('REPCOUNTRY', 'SUBAFR', 'SUBAFR', 'UNK'))
GetOriginDistribution(inputData)

}
