% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOriginGroupingPreset.R
\name{GetOriginGroupingPreset}
\alias{GetOriginGroupingPreset}
\title{GetOriginGroupingPreset}
\usage{
GetOriginGroupingPreset(type = "REPCOUNTRY + UNK + OTHER", distr)
}
\arguments{
\item{type}{Grouping type. Default = 'REPCOUNTRY + UNK + OTHER'}

\item{distr}{Distribution of RegionOfOrigin}
}
\description{
Get mapping from RegionOfOrigin to GroupOfOrigin
}
\examples{
distr <- data.table::data.table(
  origin = c('REPCOUNTRY', 'SUBAFR'),
  count = c(1536, 2237)
)
GetOriginGroupingPreset(
  type = 'REPCOUNTRY + UNK + 3 most prevalent regions + OTHER',
  distr = distr
)

}
