% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPreliminaryAttributesMapping.R
\name{GetPreliminaryAttributesMapping}
\alias{GetPreliminaryAttributesMapping}
\title{GetPreliminaryAttributesMapping}
\usage{
GetPreliminaryAttributesMapping(origData, maxDistance = 4)
}
\arguments{
\item{origData}{Original data. Required.}

\item{maxDistance}{Maximum allowed generalized Levenshtein distance between internal and origina
column name. Optional. Default = 4}
}
\value{
list with attributes mapping
}
\description{
Gets attributes mappings.
}
\examples{
\dontrun{
GetPreliminaryAttributesMapping(origData)
}

}
