% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPreliminaryDefaultValues.R
\name{GetPreliminaryDefaultValues}
\alias{GetPreliminaryDefaultValues}
\title{GetPreliminaryDefaultValues}
\usage{
GetPreliminaryDefaultValues()
}
\value{
list of default values
}
\description{
Get preliminary default values for input data columns.
}
\examples{
GetPreliminaryDefaultValues()

}
