% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRDPlotData.R
\name{GetRDPlotData}
\alias{GetRDPlotData}
\title{GetRDPlotData}
\usage{
GetRDPlotData(data, by)
}
\arguments{
\item{data}{Data table object. Required.}

\item{by}{Character vector of index columns. Required.}
}
\value{
Data table object
}
\description{
Get plot data for reporting delay adjustment artifacts.
}
\examples{
\dontrun{
GetRDPlotData(data, by)
}

}
