% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetReportFileNames.R
\name{GetReportFileNames}
\alias{GetReportFileNames}
\title{GetReportFileNames}
\usage{
GetReportFileNames(path = GetSystemFile("reports"))
}
\arguments{
\item{path}{Path to the folder with adjustment specifications. Optional.
Default = \code{\link{GetSystemFile}('reports')}.}
}
\value{
Character vector of adjustment specification file paths.
}
\description{
Get the list of report file names in a specific folder. By default the search folder is
\code{reports} folder installed with package \code{\link{hivPlatform}}.
}
\examples{
\dontrun{
GetReportFileNames()
}

}
