% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetReportingDelaysPlots.R
\name{GetReportingDelaysPlots}
\alias{GetReportingDelaysPlots}
\title{GetReportingDelaysPlots}
\usage{
GetReportingDelaysPlots(inputData)
}
\arguments{
\item{inputData}{Pre-processed input data. Required.}
}
\value{
list object
}
\description{
Create plot with overview of reporting delays data.
}
\examples{
\dontrun{
GetReportingDelaysPlots(inputData)
}

}
