% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTimeStamp.R
\name{GetTimeStamp}
\alias{GetTimeStamp}
\title{GetTimeStamp}
\usage{
GetTimeStamp(timeZone = Sys.timezone())
}
\arguments{
\item{timeZone}{Time zone to use when getting current time. Optional.
Default = \code{\link{Sys.timezone}()}.}
}
\value{
Timestamp string.
}
\description{
Get time stamp of the form "YYYYmmddHHMMSS" in a given time zone. By default the current time
zone is used.
}
\examples{
GetTimeStamp()
GetTimeStamp(timeZone = "GMT")

}
