% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hivEstimatesAccuracy2-package.R
\docType{package}
\name{hivPlatform}
\alias{hivPlatform}
\title{Improving Accuracy of HIV Estimates in EU/EEA Countries}
\description{
Improves accuracy of HIV estimates in EU/EEA countries.
}
\author{
Author: Magdalena Rosinska \email{<mrosinska@pzh.gov.pl>}\cr
Author: Nikos Pantazis \email{<npantaz@med.uoa.gr>}\cr
Author: Ard van Sighem \email{<a.i.vansighem@amsterdamumc.nl>}\cr
Creator: Daniel Lewandowski \email{<daniel@nextpagesoft.net>}
}
