% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsEmptyString.R
\name{IsEmptyString}
\alias{IsEmptyString}
\title{IsEmptyString}
\usage{
IsEmptyString(text)
}
\arguments{
\item{text}{Text to be tested. Required.}
}
\value{
string
}
\description{
Checks if the passed text is NULL or empty ("") or NA (in that order).
}
\examples{
IsEmptyString(NULL)
IsEmptyString("")
IsEmptyString(NA)
IsEmptyString(character())
IsEmptyString(5)
IsEmptyString("5")
IsEmptyString("text")

}
