% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsError.R
\name{IsError}
\alias{IsError}
\title{IsError}
\usage{
IsError(x)
}
\arguments{
\item{x}{Object. Required.}
}
\value{
logical
}
\description{
Test if is an error
}
\examples{
IsError(try({(function() stop())()}, silent = TRUE))
IsError(try({(function() 1 + 1)()}, silent = TRUE))

}
