% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.replace.R
\name{na.replace}
\alias{na.replace}
\title{na.replace}
\usage{
na.replace(x, replace)
}
\arguments{
\item{x}{A numeric vector.}

\item{replace}{Value for replacing NAs}
}
\value{
Vector of length of x, but with NA's replaced by new value.
}
\description{
Sets NAs to new value.
}
\examples{
na.replace(c(1, NA), 2)

}
