% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.zero.R
\name{na.zero}
\alias{na.zero}
\title{na.zero}
\usage{
na.zero(x)
}
\arguments{
\item{x}{A numeric vector. No test that x is numeric.}
}
\value{
Vector of length of x, but with NA's replaced by 0.
}
\description{
Sets NAs to zero.
}
\examples{
na.zero(c(1, NA))

}
