% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessInputDataBeforeAdjustments.R
\name{PreProcessInputDataBeforeAdjustments}
\alias{PreProcessInputDataBeforeAdjustments}
\title{PreProcessInputDataBeforeAdjustments}
\usage{
PreProcessInputDataBeforeAdjustments(inputData)
}
\arguments{
\item{inputData}{Input data. Required.}
}
\value{
data.table object (invisibly)
}
\description{
Pre-processes input data before passing it to adjustment scripts.
All changes are applied to the input object by reference.
There is no need to assign the result of this function.
}
\examples{
\dontrun{
PreProcessInputDataBeforeAdjustments(inputData)
}

}
