% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessInputDataBeforeSummary.R
\name{PreProcessInputDataBeforeSummary}
\alias{PreProcessInputDataBeforeSummary}
\title{PreProcessInputDataBeforeSummary}
\usage{
PreProcessInputDataBeforeSummary(inputData, seed = NULL)
}
\arguments{
\item{inputData}{Input data. Required.}

\item{seed}{Random seed. Optional. Default = NULL}
}
\value{
list object
}
\description{
Pre-processes input data before making data summary and passing it to adjustment scripts.
}
\examples{
\dontrun{
PreProcessInputDataBeforeSummary(inputData)
}

}
