% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadExcelFile.R
\name{ReadExcelFile}
\alias{ReadExcelFile}
\title{ReadExcelFile}
\usage{
ReadExcelFile(fileName, fileType, ...)
}
\arguments{
\item{fileName}{Name of file to read. Required.}

\item{fileType}{Type of data to read. Required if \code{fileName} has no file extension,
otherwise optional.}

\item{...}{Additional parameters passed to \code{\link[readxl]{read_excel}}. Optional.}
}
\value{
data.table object.
}
\description{
Read content of an Excel file.
}
\examples{
\dontrun{
ReadExcelFile(fileName)
}

}
