% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRdsFile.R
\name{ReadRdsFile}
\alias{ReadRdsFile}
\title{ReadRdsFile}
\usage{
ReadRdsFile(fileName, ...)
}
\arguments{
\item{fileName}{Name of file to read. Required.}

\item{...}{Additional parameters passed to \code{\link{readRDS}}. Optional.}
}
\value{
data.table object.
}
\description{
Read content of a rds file.
}
\examples{
\dontrun{
ReadRdsFile(fileName)
}

}
