% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRmdFrontMatter.R
\name{ReadRmdFrontMatter}
\alias{ReadRmdFrontMatter}
\title{ReadRmdFrontMatter}
\usage{
ReadRmdFrontMatter(filePath, section = NULL)
}
\arguments{
\item{filePath}{Path to the source RMarkdown file. Required.}

\item{section}{Name of section of the front matter to return. If not specified, then
the whole specification is returned, otherwise only the specified section. Optional.
Default = NULL.}
}
\value{
list object
}
\description{
Reads YAML front matter from a Rmd file.
}
\examples{
\dontrun{
ReadRmdFrontMatter(filePath)
}

}
