% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadStringFromFile.R
\name{ReadStringFromFile}
\alias{ReadStringFromFile}
\title{ReadStringFromFile}
\usage{
ReadStringFromFile(fileName)
}
\arguments{
\item{fileName}{Name of file to read. Required.}
}
\value{
String
}
\description{
Read content of a file as a string
}
\examples{
\dontrun{
ReadStringFromFile(fileName)
}

}
