% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadTextFile.R
\name{ReadTextFile}
\alias{ReadTextFile}
\title{ReadTextFile}
\usage{
ReadTextFile(fileName, ...)
}
\arguments{
\item{fileName}{Name of file to read. Required.}

\item{...}{Additional parameters passed to \code{\link[data.table]{fread}}. Optional.}
}
\value{
data.table object.
}
\description{
Read content of a (tab delimited: see parameters) text file.
}
\examples{
\dontrun{
ReadTextFile(fileName)
}

}
