% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordGgplot.R
\name{RecordGgplot}
\alias{RecordGgplot}
\title{RecordGgplot}
\usage{
RecordGgplot(plotObj)
}
\arguments{
\item{plotObj}{ggplot object}
}
\value{
recorded plot object
}
\description{
Description
}
\examples{
\dontrun{
RecordGgplot(plotObj)
}

}
