% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenderReportForAdjSpec.R
\name{RenderReportForAdjSpec}
\alias{RenderReportForAdjSpec}
\title{RenderReportForAdjSpec}
\usage{
RenderReportForAdjSpec(adjustmentSpec, fileNameSuffix = "", params = NULL)
}
\arguments{
\item{adjustmentSpec}{Adjustment specification. Required.}

\item{fileNameSuffix}{Report file suffix. Base name is taken from the adjustment spec file name.
Optional. Default = ''.}

\item{params}{Parameters passed to the report. Optional. Default = NULL.}
}
\value{
string with the rendered report
}
\description{
Renders html fragment from RMarkdown document.
}
\examples{
\dontrun{
RenderReportForAdjSpec(adjustmentSpec, fileNameSuffix, parameters)
}

}
