% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenderReportToHTML.R
\name{RenderReportToHTML}
\alias{RenderReportToHTML}
\title{RenderReportToHTML}
\usage{
RenderReportToHTML(reportFilePath, params = NULL)
}
\arguments{
\item{reportFilePath}{Path to the source RMarkdown file. Required.}

\item{params}{Parameters passed to the report. Optional. Default = NULL.}
}
\value{
string with the rendered report
}
\description{
Renders html fragment from RMarkdown document.
}
\examples{
\dontrun{
RenderReportToHTML(adjustmentSpec, fileNameSuffix, parameters)
}

}
