% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunApp.R
\name{RunApp}
\alias{RunApp}
\title{RunApp}
\usage{
RunApp(
  launchBrowser = getOption("shiny.launch.browser", interactive()),
  trace = FALSE
)
}
\arguments{
\item{launchBrowser}{Logical indicating to open the app in a newly open web browser}

\item{trace}{Logical indicating to output detailed debug info in R session}
}
\value{
NULL (invisibly)
}
\description{
Run the application.
}
\examples{
\dontrun{
RunApp()
}

}
