% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteDataFile.R
\name{WriteDataFile}
\alias{WriteDataFile}
\title{WriteDataFile}
\usage{
WriteDataFile(data, fileName, ...)
}
\arguments{
\item{data}{data.table object to save. Required.}

\item{fileName}{Name of the saved file. Required.}

\item{...}{Additional parameters passed to worker functions. Optional.}
}
\value{
NULL (invisibly)
}
\description{
Write content of a data table to file.
}
\examples{
\dontrun{
WriteDataFile(data, fileName)
}

}
