% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteStataFile.R
\name{WriteStataFile}
\alias{WriteStataFile}
\title{WriteStataFile}
\usage{
WriteStataFile(data, fileName, ...)
}
\arguments{
\item{data}{data.table object to save. Required.}

\item{fileName}{Name of the saved file. Required.}

\item{...}{Additional parameters passed to \code{\link[foreign]{write.dta}}. Optional.}
}
\value{
NULL (invisibly)
}
\description{
Write content of a data table (data frame as well) to a Stata file.
}
\examples{
\dontrun{
WriteStataFile(data, fileName)
}

}
