% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteTextFile.R
\name{WriteTextFile}
\alias{WriteTextFile}
\title{WriteTextFile}
\usage{
WriteTextFile(data, fileName, ...)
}
\arguments{
\item{data}{data.table object to save. Required.}

\item{fileName}{Name of the saved file. Required.}

\item{...}{Additional parameters passed to \code{\link[data.table]{fwrite}}. Optional.}
}
\value{
NULL (invisibly)
}
\description{
Write content of a data table (data frame as well) to a text file.
}
\examples{
\dontrun{
WriteTextFile(data, fileName)
}

}
