% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteZipFile.R
\name{WriteZipFile}
\alias{WriteZipFile}
\title{WriteZipFile}
\usage{
WriteZipFile(hivModelData, outputFilePath = NULL, ...)
}
\arguments{
\item{hivModelData}{List of data.table objects. Required.}

\item{outputFilePath}{Path to the output file. Required. Default = \code{NULL}.}

\item{...}{Additional arguments passed to the internal writing function.}
}
\value{
File name of the saved zip file
}
\description{
Saves sets of files to a zip file. The file contains multiple folders, one for each
member of input list.
}
\examples{
\dontrun{
WriteZipFile(hivModelData)
}

}
