% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAvailableStrata.R
\name{GetAvailableStrata}
\alias{GetAvailableStrata}
\title{GetAvailableStrata}
\usage{
GetAvailableStrata(
  dt,
  variables = c(Gender = "G", Transmission = "T", GroupedRegionOfOrigin = "O",
    PlaceOfResidence = "R")
)
}
\arguments{
\item{dt}{dt}

\item{variables}{variables}
}
\value{
list
}
\description{
Get available strata
}
\examples{
dt <- data.table::data.table(
  Imputation = c(0, 1, 2),
  Gender = c('F', 'F', 'M'),
  Transmission = c('IDU', 'IDU', 'MSM')
)
GetAvailableStrata(dt, variables = c('Gender' = 'G', 'Transmission' = 'T'))

}
