% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSystemFile.R
\name{GetSystemFile}
\alias{GetSystemFile}
\title{GetSystemFile}
\usage{
GetSystemFile(..., package = "hivEstimatesAccuracy2")
}
\arguments{
\item{...}{Arguments for \code{\link{system.file}}}

\item{package}{Package name. Default = \code{'hivEstimatesAccuracy2'}.}
}
\value{
Path to a system file
}
\description{
Get the system path - 'inst' folder.
}
\examples{
GetSystemFile('adjustments')

}
